+++ src/Opulence/Authentication/Users/User.php
@@ -6,9 +6,9 @@
  * @copyright Copyright (C) 2016 David Young
  * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
  */
-namespace Opulence\Users;
+namespace Opulence\Authentication\Users;
 
-use DateTime;
+use DateTimeImmutable;
 
 /**
  * Defines a user
@@ -17,19 +17,23 @@ class User implements IUser
 {
     /** @var int The database Id of the user */
     protected $id = -1;
-    /** @var DateTime The date this user was created */
+    /** @var string The hashed password of the user */
+    protected $hashedPassword = "";
+    /** @var DateTimeImmutable The date this user was created */
     protected $dateCreated = null;
     /** @var array The list of roles this user has */
     protected $roles = [];
 
     /**
      * @param int $id The database Id of this user
-     * @param DateTime $dateCreated The date this user was created
+     * @param string $hashedPassword The hashed password of this user
+     * @param DateTimeImmutable $dateCreated The date this user was created
      * @param array $roles The list of roles this user has
      */
-    public function __construct($id, DateTime $dateCreated, array $roles = [])
+    public function __construct($id, $hashedPassword, DateTimeImmutable $dateCreated, array $roles = [])
     {
         $this->setId($id);
+        $this->setHashedPassword($hashedPassword);
         $this->dateCreated = $dateCreated;
         $this->roles = $roles;
     }
@@ -40,6 +44,14 @@ class User implements IUser
     public function getDateCreated()
     {
         return $this->dateCreated;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getHashedPassword()
+    {
+        return $this->hashedPassword;
     }
 
     /**
@@ -64,6 +76,14 @@ class User implements IUser
     public function hasRole($role)
     {
         return in_array($role, $this->roles);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setHashedPassword($hashedPassword)
+    {
+        $this->hashedPassword = $hashedPassword;
     }
 
     /**