+++ src/Opulence/Authentication/Users/UserRoles.php
@@ -6,7 +6,7 @@
  * @copyright Copyright (C) 2016 David Young
  * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
  */
-namespace Opulence\Users;
+namespace Opulence\Authentication\Users;
 
 /**
  * Defines the different user role types