+++ tests/src/Opulence/Authentication/Tokens/TokenTest.php
@@ -8,7 +8,7 @@
  */
 namespace Opulence\Authentication\Tokens;
 
-use DateTime;
+use DateTimeImmutable;
 use Opulence\Tests\Authentication\Tokens\Mocks\Token;
 
 /**
@@ -21,9 +21,9 @@ class TokenTest extends \PHPUnit_Framework_TestCase
      */
     public function testCheckingIsActiveWithFutureValidFrom()
     {
-        $validFrom = new DateTime("+1 day");
-        $validTo = new DateTime("+1 week");
-        $token = new Token(1, "", $validFrom, $validTo, true);
+        $validFrom = new DateTimeImmutable("+1 day");
+        $validTo = new DateTimeImmutable("+1 week");
+        $token = new Token(1, 2, "", $validFrom, $validTo, true);
         $this->assertFalse($token->isActive());
     }
 
@@ -32,9 +32,9 @@ class TokenTest extends \PHPUnit_Framework_TestCase
      */
     public function testCheckingIsActiveWithFutureValidTo()
     {
-        $validFrom = new DateTime("now");
-        $validTo = new DateTime("+1 week");
-        $token = new Token(1, "", $validFrom, $validTo, true);
+        $validFrom = new DateTimeImmutable("now");
+        $validTo = new DateTimeImmutable("+1 week");
+        $token = new Token(1, 2, "", $validFrom, $validTo, true);
         $this->assertTrue($token->isActive());
     }
 
@@ -43,9 +43,9 @@ class TokenTest extends \PHPUnit_Framework_TestCase
      */
     public function testCheckingIsActiveWithInactiveToken()
     {
-        $validFrom = new DateTime("-1 week");
-        $validTo = new DateTime("+1 week");
-        $token = new Token(1, "", $validFrom, $validTo, false);
+        $validFrom = new DateTimeImmutable("-1 week");
+        $validTo = new DateTimeImmutable("+1 week");
+        $token = new Token(1, 2, "", $validFrom, $validTo, false);
         $this->assertFalse($token->isActive());
     }
 
@@ -54,9 +54,9 @@ class TokenTest extends \PHPUnit_Framework_TestCase
      */
     public function testCheckingIsActiveWithPastValidFrom()
     {
-        $validFrom = new DateTime("-1 week");
-        $validTo = new DateTime("+1 week");
-        $token = new Token(1, "", $validFrom, $validTo, true);
+        $validFrom = new DateTimeImmutable("-1 week");
+        $validTo = new DateTimeImmutable("+1 week");
+        $token = new Token(1, 2, "", $validFrom, $validTo, true);
         $this->assertTrue($token->isActive());
     }
 
@@ -65,9 +65,9 @@ class TokenTest extends \PHPUnit_Framework_TestCase
      */
     public function testCheckingIsActiveWithPastValidTo()
     {
-        $validFrom = new DateTime("now");
-        $validTo = new DateTime("-1 week");
-        $token = new Token(1, "", $validFrom, $validTo, true);
+        $validFrom = new DateTimeImmutable("now");
+        $validTo = new DateTimeImmutable("-1 week");
+        $token = new Token(1, 2, "", $validFrom, $validTo, true);
         $this->assertFalse($token->isActive());
     }
 
@@ -77,9 +77,9 @@ class TokenTest extends \PHPUnit_Framework_TestCase
     public function testGettingHashedValue()
     {
         $hashedValue = "foo";
-        $validFrom = new DateTime("-1 week");
-        $validTo = new DateTime("+1 week");
-        $token = new Token(1, $hashedValue, $validFrom, $validTo, false);
+        $validFrom = new DateTimeImmutable("-1 week");
+        $validTo = new DateTimeImmutable("+1 week");
+        $token = new Token(1, 2, $hashedValue, $validFrom, $validTo, false);
         $this->assertEquals($hashedValue, $token->getHashedValue());
     }
 
@@ -89,10 +89,21 @@ class TokenTest extends \PHPUnit_Framework_TestCase
     public function testGettingId()
     {
         $id = 1;
-        $validFrom = new DateTime("-1 week");
-        $validTo = new DateTime("+1 week");
-        $token = new Token($id, "", $validFrom, $validTo, true);
+        $validFrom = new DateTimeImmutable("-1 week");
+        $validTo = new DateTimeImmutable("+1 week");
+        $token = new Token($id, 2, "", $validFrom, $validTo, true);
         $this->assertEquals($id, $token->getId());
+    }
+
+    /**
+     * Tests getting the user Id
+     */
+    public function testGettingUserId()
+    {
+        $validFrom = new DateTimeImmutable("-1 week");
+        $validTo = new DateTimeImmutable("+1 week");
+        $token = new Token(1, 2, "", $validFrom, $validTo, true);
+        $this->assertEquals(2, $token->getUserId());
     }
 
     /**
@@ -100,8 +111,8 @@ class TokenTest extends \PHPUnit_Framework_TestCase
      */
     public function testGettingValidFromDate()
     {
-        $validFromDate = new DateTime("1776-07-04 12:34:56");
-        $token = new Token(1, "", $validFromDate, new DateTime("1970-01-01 01:00:00"), true);
+        $validFromDate = new DateTimeImmutable("1776-07-04 12:34:56");
+        $token = new Token(1, 2, "", $validFromDate, new DateTimeImmutable("1970-01-01 01:00:00"), true);
         $this->assertEquals($validFromDate, $token->getValidFrom());
     }
 
@@ -110,8 +121,8 @@ class TokenTest extends \PHPUnit_Framework_TestCase
      */
     public function testGettingValidToDate()
     {
-        $validToDate = new DateTime("1970-01-01 01:00:00");
-        $token = new Token(1, "", new DateTime("1776-07-04 12:34:56"), $validToDate, true);
+        $validToDate = new DateTimeImmutable("1970-01-01 01:00:00");
+        $token = new Token(1, 2, "", new DateTimeImmutable("1776-07-04 12:34:56"), $validToDate, true);
         $this->assertEquals($validToDate, $token->getValidTo());
     }
 
@@ -122,9 +133,9 @@ class TokenTest extends \PHPUnit_Framework_TestCase
     {
         $oldId = 1;
         $newId = 2;
-        $validFrom = new DateTime("-1 week");
-        $validTo = new DateTime("+1 week");
-        $token = new Token($oldId, "", $validFrom, $validTo, true);
+        $validFrom = new DateTimeImmutable("-1 week");
+        $validTo = new DateTimeImmutable("+1 week");
+        $token = new Token($oldId, 2, "", $validFrom, $validTo, true);
         $token->setId($newId);
         $this->assertEquals($newId, $token->getId());
     }